Refactor useEffect dependency and update LanguageSwitch
In ArticlesPage, useEffect dependency has been refactored to focus on the whole 'sanitizedArticlesData' instead of its length. In LanguageSwitch, navigation behavior was added to improve user interaction. It not only changes the language but navigates to the home page as well. Also, a styling update was made for better responsiveness on different screen sizes.